Added scratchpad directory
This is a temporary and experimental branch related to the swt-os2
project at GitHub. So there'll be some fiddling and to not pollute the
normal branch structure, wel'll do all that in the scratchpad directory.

'$' as a name is short, appears at the top in FC/2 panels and dir
listings and indicates something discardable.

Since Git does not store empty directories, we'll put a dummy file
inside it.

A side effect when using git-svn is that all git commits made at the
dev-side are rewritten before they are committed to the svn repository
it was cloned from. This means that all these commits have their hashes
and commit-time changed and appear at the svn side as one bulk commit.

git-svn-id: http://svn.netlabs.org/repos/odin32/branches/swt@22069 3ae8c665-79e0-dc11-ae90-00e081727c95