Fix crash in exception handling when using JWasm
When externs are declared in the code-segment, JWasm generates
fixups relative to that segment. Because CODE32 is not part of CGROUP,
and JWasm does not treat FLAT as special, these fixups are incorrect.

One way would be to move the extern declarations outside of the
code-segment, but then the symantic difference between JWasm and Alp
would remain.

A better solution is to make CODE32 part of CGROUP, which causes JWasm
to generate the correct GROUP relative fixups.

This issue applies to linking object modules generated with GCC and
regular PC tools, which introduce TEXT32 and CODE32 segments. Grouping
them ensures group relative fixups are calculated.

git-svn-id: http://svn.netlabs.org/repos/odin32/branches/swt@22087 3ae8c665-79e0-dc11-ae90-00e081727c95