Changed application termination                                [apitest]
Termination is now redirected to WM_CLOSE.
WM_CLOSE does a DestroyWindow(), which in turn generates a WM_DESTROY
message, which is where the main message-loop is ended by doing a call
to PostQuitMessage().

At exit however, the 'wParam' field of the 'msg' structure does not
contain the exit-code as provided with PostQuitMessage().

git-svn-id: http://svn.netlabs.org/repos/odin32/branches/swt@22115 3ae8c665-79e0-dc11-ae90-00e081727c95